[CHG][N/T] 1. 修复姿态信息被截断的问题 2. 修复姿态信息横摇和纵摇反向的问题 3. 修复姿态信息 float 被取整截断的问题

+++ led3000/network_protocol.cpp
@@ -107,7 +107,7 @@ static int do_with_network_attitude_info(NetworkPackage &net_package)
 
     memcpy(&vertical_info, net_package.payload() + 6, sizeof(vertical_info));
     vertical_info = (int)ntohl(vertical_info);
-    vertical_info_float = (int)_do_format_dev_int2float(vertical_info, DIMENSION_90_INT);
+    vertical_info_float = _do_format_dev_int2float(vertical_info, DIMENSION_90_INT);
 
     memcpy(&horizon_info, net_package.payload() + 10, sizeof(horizon_info));
     horizon_info = (int)ntohl(horizon_info);
